requirements.txt
alabaster==0.7.8
apt-clone==0.2.1
apturl==0.5.2
astroid==1.6.0
Babel==2.4.0
beautifulsoup4==4.6.0
bleach==2.1.2
Brlapi==0.6.6
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
cloudpickle==0.5.2
command-not-found==0.3
configobj==5.0.6
cupshelpers==1.0
cycler==0.10.0
Cython==0.26.1
decorator==4.1.2
defer==1.0.6
docutils==0.14
entrypoints==0.2.3.post1
Flask==1.1.2
Flask-SQLAlchemy==2.4.3
Guake==3.0.5
h5py==2.7.1
html5lib==0.999999999
httplib2==0.9.2
idna==2.9
imagesize==0.7.1
iotop==0.6
ipykernel==4.8.2
ipython==5.5.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.11.1
Jinja2==2.11.2
joblib==0.15.1
jsonschema==2.6.0
jupyter-client==5.2.2
jupyter-core==4.4.0
kiwisolver==1.2.0
lazy-object-proxy==1.3.1
leveldb==0.1
logilab-common==1.4.1
louis==3.5.0
lxml==4.2.4
macaroonbakery==1.1.3
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nemo-emblems==4.4.0
netifaces==0.10.4
networkx==1.11
nose==1.3.7
numpy==1.18.5
numpydoc==0.7.0
onboard==1.4.1
opencv-python==4.2.0.32
PAM==0.4.2
pandas==1.0.4
pandocfilters==1.4.2
parso==0.1.1
pbr==3.1.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==7.0.0
prompt-toolkit==1.0.15
protobuf==3.11.2
psutil==5.4.2
pycairo==1.16.2
pycodestyle==2.3.1
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pygobject==3.26.1
PyICU==1.9.8
pyinotify==0.9.6
pylint==1.8.3
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyparted==3.11.1
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-gflags==1.5.1
python-xapp==1.8.1
python-xlib==0.20
pytz==2020.1
PyWavelets==0.5.1
pyxdg==0.25
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
reportlab==3.4.0
requests==2.23.0
requests-unixsocket==0.1.5
roman==2.0.0
rope==0.10.5
scikit-image==0.13.1
scikit-learn==0.23.1
scipy==1.4.1
sessioninstaller==0.0.0
setproctitle==1.1.10
simplegeneric==0.8.1
sip==4.19.8
six==1.15.0
sklearn==0.0
Sphinx==1.6.7
spyder==3.2.6
SQLAlchemy==1.3.17
system-service==0.3
systemd-python==234
testpath==0.3.1
threadpoolctl==2.1.0
tinycss==0.4
tornado==4.5.3
traitlets==4.3.2
ubuntu-drivers-common==0.0.0
ufw==0.36
urllib3==1.25.9
virtualenv==15.1.0
wcwidth==0.1.7
webencodings==0.5
Werkzeug==1.0.1
wfdb==2.2.1
wrapt==1.9.0
xkit==0.0.0
